requirements.txt
Django==5.0.6
django-extensions==3.2.3
djangorestframework==3.15.2
gunicorn==22.0.0
keras==3.3.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
xgboost==2.1.0
